requirements.txt
aioredis==1.3.1
amqp==5.0.9
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==21.12b0
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Deprecated==1.2.13
Django==4.0
django-admin-inline-paginator==0.3.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.1
Faker==11.1.0
flake8==4.0.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.10.0
incremental==21.3.0
iniconfig==1.1.1
kombu==5.2.3
Markdown==3.3.6
mccabe==0.6.1
msgpack==1.0.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.24
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyotp==2.6.0
pyparsing==3.0.6
pytest==6.2.5
pytest-asyncio==0.16.0
pytest-django==4.5.2
pytest-mock==3.6.1
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.7.0
zope.interface==5.4.0
